First version of the secondary structure analysis for ModelMaker
It is now printing a plot with the prevalence of ss for each residue
using multiplot and exporting to a postscript file.

ISSUE: the background of the plot is not exported to the ps file
The ResID categories in the plot is still a floating point number, so depending
on the number of residues, it might show as a fractional number.

Still, need to develop the searching for different ss patterns